feat(sidecar): update primitives
